Prevent callback being called twice when it throws an error
When the callback threw an error it got caught in catch block which
called the callback again but now with the error. This resulted in the
callback being called at least twice, and potentially leading to infinite loops.
By extracting the call to the callback to outside the try-catch block this is
prevented